JWK as injected dependency
